Fixing imports so first trident import and trident.verify both work.
